+++ packages/ui/src/index.ts
@@ -1,2 +1,3 @@
 export * from './layout/app-layout'
+export * from './menu'
 export * from './messages'